Update layout metadata and add providers
