+++ app/app/(appmap)/layout.tsx
@@ -1,6 +1,13 @@
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
+import '@/styles/globals.css'
+import type { Metadata } from 'next'
+import { Inter } from 'next/font/google'
+import { Providers } from '@/app/providers'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata: Metadata = {
+  title: 'Map',
+  description: 'What&apos;s around campus?',
 }
 
 export default function RootLayout({
@@ -10,7 +17,11 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body className={inter.className}>
+        <Providers>
+          {children}
+        </Providers>
+      </body>
     </html>
   )
 }
