--- app/app/(appmap)/map/marker/page.tsx
@@ -1,7 +1,0 @@
-export default function Page() {
-  return (
-    <main>
-      <h1> My Markers </h1>
-    </main>
-  )
-}
